symbiosis/pyproject.toml
[tool.poetry]
name = "symbiosis"
version = "0.1.0"
description = "Symbiosis"
authors = ["Debajyoti Raychaudhuri <draychaudhuri1086@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.8"
enduro-bot = { path = "enduro-bot/" }
flappy-bird-bot = { path = "flappy-bird-bot/" }
flappy-bird-with-gunner-farneback-bot = { path = "flappy-bird-with-gunner-farneback-bot/" }

[tool.poetry.dev-dependencies]
tensorflow-gpu = "2.2.0"
opencv-python = "^4.4.0.44"
dill = "^0.3.2"
loguru = "^0.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
